projects/edge-dev-main-backup-20260126/backend/lc_d2/verify_accuracy.py
"""
Verify LC D2 scanner accuracy against original
Tests only the tickers that passed validation in fixed_formatted.py
"""
import pandas as pd
import numpy as np
from dotenv import load_dotenv
import os
import sys

# Import functions from source.py
sys.path.insert(0, '/Users/michaeldurante/ai dev/ce-hub/projects/edge-dev-exact/templates/lc_d2')
from source import (
    fetch_data_multi,
    compute_atr,
    compute_emas,
    compute_higher_low,
    check_high_lvl_filter_lc,
    get_trading_date_offsets,
    check_lc_pm_liquidity,
    get_min_price_lc,
    check_next_day_valid_lc,
    filter_lc_rows
)

load_dotenv()

# Load fixed_formatted results
fixed_results = pd.read_csv('lc_d2_results.csv')
fixed_results['Date'] = pd.to_datetime(fixed_results['Date']).dt.date

print(f"Fixed formatted scanner found {len(fixed_results)} validated signals")
print(f"\nSignals to verify:")
print(fixed_results.to_string(index=False))

# Get unique tickers and dates
unique_tickers = fixed_results['Ticker'].unique().tolist()
unique_dates = fixed_results['Date'].unique().tolist()

print(f"\n{'='*70}")
print(f"Testing {len(unique_tickers)} tickers on {len(unique_dates)} dates")
print(f"{'='*70}")

# Fetch data for these tickers
START_DATE = '2024-11-01'  # Need enough history for calculations
END_DATE = '2025-12-31'

print(f"\nFetching data from {START_DATE} to {END_DATE}...")

try:
    df = fetch_data_multi(unique_tickers, START_DATE, END_DATE)

    if df.empty:
        print("❌ No data fetched!")
        sys.exit(1)

    print(f"✅ Fetched {len(df):,} rows")

    # Compute features (matching original pipeline)
    print("\nComputing features...")
    df = compute_atr(df)
    df = compute_emas(df)
    df = compute_higher_lower(df)

    # Add trading date offsets
    df = get_trading_date_offsets(df)

    # Check PM liquidity
    print("Checking PM liquidity...")
    df = check_lc_pm_liquidity(df)

    # Apply LC pattern detection
    print("Detecting patterns...")
    df = check_high_lvl_filter_lc(df)

    # Filter to signals
    df = filter_lc_rows(df)

    # Get min prices (for completeness, though not used in validation due to bug)
    df = get_min_price_lc(df)

    # Check next day validity (won't work due to bug in original - min_price created after this call)
    # df = check_next_day_valid_lc(df)  # Skip this as it's broken in original

    # Filter to our target dates and tickers
    original_results = df[
        (df['ticker'].isin(unique_tickers)) &
        (df['date'].isin(unique_dates))
    ].copy()

    # Get pattern columns
    pattern_cols = ['lc_frontside_d3_extended_1', 'lc_frontside_d2_extended', 'lc_frontside_d2_extended_1']

    # Create results from original
    orig_signals = []
    for _, row in original_results.iterrows():
        patterns = []
        for col in pattern_cols:
            if col in df.columns and row.get(col) == 1:
                patterns.append(col)

        if patterns:
            orig_signals.append({
                'Ticker': row['ticker'],
                'Date': row['date'],
                'Scanner_Label': ', '.join(patterns)
            })

    if orig_signals:
        orig_df = pd.DataFrame(orig_signals)
        print(f"\n{'='*70}")
        print(f"Original scanner found {len(orig_df)} signals on these dates/tickers")
        print(f"{'='*70}")
        print(orig_df.to_string(index=False))

        # Compare results
        comparison = []
        for _, fixed_row in fixed_results.iterrows():
            ticker = fixed_row['Ticker']
            date = fixed_row['Date']
            fixed_patterns = set(fixed_row['Scanner_Label'].split(', '))

            # Find matching original signal
            orig_match = orig_df[
                (orig_df['Ticker'] == ticker) &
                (orig_df['Date'] == date)
            ]

            if not orig_match.empty:
                orig_patterns = set(orig_match.iloc[0]['Scanner_Label'].split(', '))

                # Check if patterns match
                if fixed_patterns == orig_patterns:
                    comparison.append({
                        'Ticker': ticker,
                        'Date': date,
                        'Status': '✅ MATCH',
                        'Fixed_Patterns': ', '.join(sorted(fixed_patterns)),
                        'Orig_Patterns': ', '.join(sorted(orig_patterns))
                    })
                else:
                    comparison.append({
                        'Ticker': ticker,
                        'Date': date,
                        'Status': '⚠️  PATTERN MISMATCH',
                        'Fixed_Patterns': ', '.join(sorted(fixed_patterns)),
                        'Orig_Patterns': ', '.join(sorted(orig_patterns))
                    })
            else:
                comparison.append({
                    'Ticker': ticker,
                    'Date': date,
                    'Status': '❌ NOT IN ORIGINAL',
                    'Fixed_Patterns': ', '.join(sorted(fixed_patterns)),
                    'Orig_Patterns': 'N/A'
                })

        # Check for signals in original but not in fixed
        for _, orig_row in orig_df.iterrows():
            ticker = orig_row['Ticker']
            date = orig_row['Date']
            orig_patterns = set(orig_row['Scanner_Label'].split(', '))

            fixed_match = fixed_results[
                (fixed_results['Ticker'] == ticker) &
                (fixed_results['Date'] == date)
            ]

            if fixed_match.empty:
                comparison.append({
                    'Ticker': ticker,
                    'Date': date,
                    'Status': '⚠️  ONLY IN ORIGINAL',
                    'Fixed_Patterns': 'N/A',
                    'Orig_Patterns': ', '.join(sorted(orig_patterns))
                })

        comp_df = pd.DataFrame(comparison)
        print(f"\n{'='*70}")
        print(f"COMPARISON RESULTS")
        print(f"{'='*70}")
        print(comp_df.to_string(index=False))

        # Summary
        match_count = (comp_df['Status'] == '✅ MATCH').sum()
        total_count = len(comp_df)

        print(f"\n{'='*70}")
        print(f"ACCURACY: {match_count}/{total_count} signals match ({100*match_count/total_count:.1f}%)")
        print(f"{'='*70}")

        if match_count == total_count:
            print("✅ 100% ACCURACY CONFIRMED!")
        else:
            print("⚠️  Some discrepancies found - review above")

    else:
        print("\n❌ Original scanner found no signals on these dates/tickers")
        print("This could indicate a problem with the original scanner or the test data")

except Exception as e:
    print(f"\n❌ Error during verification: {e}")
    import traceback
    traceback.print_exc()
