projects/edge-dev-main-backup-20260126/backend/requirements.txt
# FastAPI Backend Requirements
# Preserves all dependencies from original scanner while adding FastAPI

# FastAPI and web server - Updated for Python 3.13
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Rate limiting and concurrency control
slowapi==0.1.9
redis==5.0.1

# Data processing (from original scanner) - Python 3.13 compatible versions
pandas>=2.2.0
numpy>=1.26.0
pandas-market-calendars>=4.4.0

# HTTP client (from original scanner)
aiohttp==3.9.1
requests==2.31.0

# Threading and multiprocessing (built-in, but listing for clarity)
# concurrent.futures (built-in)
# multiprocessing (built-in)
# threading (built-in)
# asyncio (built-in)

# Data processing and analysis (from original scanner)
dask[complete]==2023.12.0
dask[distributed]==2023.12.0

# Visualization (from original scanner)
plotly==5.17.0
tabulate==0.9.0

# Utilities (from original scanner)
backoff==2.2.1

# Pydantic for data validation
pydantic==2.5.0

# Logging and monitoring
structlog==23.2.0

# System monitoring
psutil==5.9.6

# Environment management
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: Enhanced monitoring
prometheus-client==0.19.0

# Optional: Better error handling
sentry-sdk[fastapi]==1.38.0